refactor: rename a few funcs to set/get parameter to the context
